added generic querier,msg parser and mock
