+++ packages/desmos-std/src/lib.rs
@@ -1,2 +1,7 @@
+pub mod msg;
 pub mod profiles;
+pub mod query;
 pub mod types;
+
+#[cfg(not(target_arch = "wasm32"))]
+pub mod mock;
