+++ packages/desmos-std/src/profiles/mod.rs
@@ -1,6 +1,4 @@
-#[cfg(not(target_arch = "wasm32"))]
-pub mod mocks;
-
+pub mod mock;
 pub mod models_app_links;
 pub mod models_blocks;
 pub mod models_chain_links;
@@ -9,7 +7,7 @@ pub mod models_dtag_requests;
 pub mod models_profile;
 pub mod models_query;
 pub mod models_relationships;
-pub mod msg_router;
-pub mod msg_routes;
+pub mod msg;
+pub mod msg_builder;
 pub mod querier;
 pub mod query_router;
