+++ packages/desmos-std/src/profiles/msg.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
-pub enum ProfilesMsgs {
+pub enum ProfilesMsg {
     SaveProfile {
         dtag: String,
         nickname: String,
