+++ packages/desmos-std/src/profiles/querier.rs
@@ -6,27 +6,28 @@ use crate::{
             QueryProfileResponse, QueryRelationshipsResponse, QueryUserApplicationLinkResponse,
             QueryUserChainLinkResponse,
         },
-        query_router::{ProfilesQueryRoute, ProfilesQueryRouter},
+        query_router::ProfilesQuery,
     },
+    query::{DesmosQuery, DesmosQueryRouter},
     types::{DesmosRoute, PageRequest},
 };
 use cosmwasm_std::{Addr, Querier, QuerierWrapper, StdResult};
 
 pub struct ProfilesQuerier<'a> {
-    querier: QuerierWrapper<'a, ProfilesQueryRouter>,
+    querier: QuerierWrapper<'a, DesmosQueryRouter>,
 }
 
 impl<'a> ProfilesQuerier<'a> {
     pub fn new(querier: &'a dyn Querier) -> Self {
         Self {
-            querier: QuerierWrapper::<'a, ProfilesQueryRouter>::new(querier),
+            querier: QuerierWrapper::<'a, DesmosQueryRouter>::new(querier),
         }
     }
 
     pub fn query_profile(&self, user: Addr) -> StdResult<QueryProfileResponse> {
-        let request = ProfilesQueryRouter {
+        let request = DesmosQueryRouter {
             route: DesmosRoute::Profiles,
-            query_data: ProfilesQueryRoute::Profile { user },
+            query_data: DesmosQuery::Profiles(ProfilesQuery::Profile { user }),
         };
 
         let res: QueryProfileResponse = self.querier.query(&request.into())?;
@@ -39,13 +40,13 @@ impl<'a> ProfilesQuerier<'a> {
         subspace_id: u64,
         pagination: Option<PageRequest>,
     ) -> StdResult<QueryRelationshipsResponse> {
-        let request = ProfilesQueryRouter {
+        let request = DesmosQueryRouter {
             route: DesmosRoute::Profiles,
-            query_data: ProfilesQueryRoute::Relationships {
+            query_data: DesmosQuery::Profiles(ProfilesQuery::Relationships {
                 user,
                 subspace_id,
                 pagination,
-            },
+            }),
         };
 
         let res: QueryRelationshipsResponse = self.querier.query(&request.into())?;
@@ -57,12 +58,12 @@ impl<'a> ProfilesQuerier<'a> {
         receiver: Addr,
         pagination: Option<PageRequest>,
     ) -> StdResult<QueryIncomingDtagTransferRequestResponse> {
-        let request = ProfilesQueryRouter {
+        let request = DesmosQueryRouter {
             route: DesmosRoute::Profiles,
-            query_data: ProfilesQueryRoute::IncomingDtagTransferRequests {
+            query_data: DesmosQuery::Profiles(ProfilesQuery::IncomingDtagTransferRequests {
                 receiver,
                 pagination,
-            },
+            }),
         };
 
         let res: QueryIncomingDtagTransferRequestResponse = self.querier.query(&request.into())?;
@@ -75,13 +76,13 @@ impl<'a> ProfilesQuerier<'a> {
         subspace_id: u64,
         pagination: Option<PageRequest>,
     ) -> StdResult<QueryBlocksResponse> {
-        let request = ProfilesQueryRouter {
+        let request = DesmosQueryRouter {
             route: DesmosRoute::Profiles,
-            query_data: ProfilesQueryRoute::Blocks {
+            query_data: DesmosQuery::Profiles(ProfilesQuery::Blocks {
                 user,
                 subspace_id,
                 pagination,
-            },
+            }),
         };
 
         let res: QueryBlocksResponse = self.querier.query(&request.into())?;
@@ -93,9 +94,9 @@ impl<'a> ProfilesQuerier<'a> {
         user: Addr,
         pagination: Option<PageRequest>,
     ) -> StdResult<QueryChainLinksResponse> {
-        let request = ProfilesQueryRouter {
+        let request = DesmosQueryRouter {
             route: DesmosRoute::Profiles,
-            query_data: ProfilesQueryRoute::ChainLinks { user, pagination },
+            query_data: DesmosQuery::Profiles(ProfilesQuery::ChainLinks { user, pagination }),
         };
 
         let res: QueryChainLinksResponse = self.querier.query(&request.into())?;
@@ -108,13 +109,13 @@ impl<'a> ProfilesQuerier<'a> {
         chain_name: String,
         target: String,
     ) -> StdResult<QueryUserChainLinkResponse> {
-        let request = ProfilesQueryRouter {
+        let request = DesmosQueryRouter {
             route: DesmosRoute::Profiles,
-            query_data: ProfilesQueryRoute::UserChainLink {
+            query_data: DesmosQuery::Profiles(ProfilesQuery::UserChainLink {
                 user,
                 chain_name,
                 target,
-            },
+            }),
         };
 
         let res: QueryUserChainLinkResponse = self.querier.query(&request.into())?;
@@ -126,9 +127,9 @@ impl<'a> ProfilesQuerier<'a> {
         user: Addr,
         pagination: Option<PageRequest>,
     ) -> StdResult<QueryApplicationLinksResponse> {
-        let request = ProfilesQueryRouter {
+        let request = DesmosQueryRouter {
             route: DesmosRoute::Profiles,
-            query_data: ProfilesQueryRoute::AppLinks { user, pagination },
+            query_data: DesmosQuery::Profiles(ProfilesQuery::AppLinks { user, pagination }),
         };
 
         let res: QueryApplicationLinksResponse = self.querier.query(&request.into())?;
@@ -141,13 +142,13 @@ impl<'a> ProfilesQuerier<'a> {
         application: String,
         username: String,
     ) -> StdResult<QueryUserApplicationLinkResponse> {
-        let request = ProfilesQueryRouter {
+        let request = DesmosQueryRouter {
             route: DesmosRoute::Profiles,
-            query_data: ProfilesQueryRoute::UserAppLinks {
+            query_data: DesmosQuery::Profiles(ProfilesQuery::UserAppLinks {
                 user,
                 application,
                 username,
-            },
+            }),
         };
 
         let res: QueryUserApplicationLinkResponse = self.querier.query(&request.into())?;
@@ -158,9 +159,11 @@ impl<'a> ProfilesQuerier<'a> {
         &self,
         client_id: String,
     ) -> StdResult<QueryApplicationLinkByClientIDResponse> {
-        let request = ProfilesQueryRouter {
+        let request = DesmosQueryRouter {
             route: DesmosRoute::Profiles,
-            query_data: ProfilesQueryRoute::ApplicationLinkByChainID { client_id },
+            query_data: DesmosQuery::Profiles(ProfilesQuery::ApplicationLinkByChainID {
+                client_id,
+            }),
         };
 
         let res: QueryApplicationLinkByClientIDResponse = self.querier.query(&request.into())?;
